Revert "Refactor bot"
This reverts commit 7f36a22349b90cdc3eb3b0ca7179cb4a2299aea1.

+++ cogs/tickets.py
@@ -1,19 +1,15 @@
 from __future__ import annotations
-
-from typing import TYPE_CHECKING, Any, Dict, Optional, Protocol, TypedDict
-
-import utils
+from typing import TYPE_CHECKING, Protocol, Any, TypedDict, Optional
 
 if TYPE_CHECKING:
-    from asyncpg import Pool
-
     from ..main import JDBot, JDBotContext
+    from asyncpg import Pool
 
     class RecordType(Protocol):
         def get(self, key: str) -> Any:
             ...
 
-    class TicketType(TypedDict):
+    class TicketCacheData(TypedDict):
         author: discord.User
         remote: int
         timestamp: datetime.datetime
@@ -21,29 +17,19 @@ if TYPE_CHECKING:
 
 import datetime
 import time
-import traceback
-from dataclasses import dataclass
 
 import discord
 from discord.ext import commands
+import traceback
 
 
-@dataclass
-class Ticket:
-    author: int
-    remote: int
-    timestamp: datetime.datetime
-
-
-class TicketCog(commands.Cog, name="Ticket"):
+class Ticket(commands.Cog):
     def __init__(self, bot: JDBot):
         self.bot = bot
-        self.pool: Optional[Pool] = self.bot.db
-        self.ticket_cache: Dict[int, Ticket] = {}
-        self.main_channel: Optional[
-            discord.TextChannel
-        ] = None  # filled in create_ticket
-        self.support_role: Optional[discord.Role] = None  # filled in create_ticket
+        self.pool: Pool = self.bot.db
+        self.ticket_cache: dict[int, TicketCacheData] = {}
+        self.main_channel: discord.TextChannel = None  # filled in create_ticket
+        self.support_role: discord.Role = None  # filled in create_ticket
 
     async def cog_load(self):
         pool = self.pool
@@ -55,70 +41,56 @@ class TicketCog(commands.Cog, name="Ticket"):
             remote: int = record.get("remote_id")
             end_timestamp: int = (record.get("end_timestamp")) / 1000
 
-            end_timestamp = datetime.datetime.fromtimestamp(
-                end_timestamp, tz=datetime.timezone.utc
-            )  # type: ignore
-            if now > end_timestamp:  # type: ignore
+            end_timestamp = datetime.datetime.fromtimestamp(end_timestamp, tz=datetime.timezone.utc)
+            if now > end_timestamp:
                 # How did this even exist in first place ?
-                await pool.execute(
-                    "DELETE FROM TICKETS WHERE author_id = $1 AND remote_id = $2",
-                    author,
-                    remote,
-                )
+                await pool.execute("DELETE FROM TICKETS WHERE author_id = $1 AND remote_id = $2", author, remote)
                 continue
-            self.ticket_cache[author] = self.ticket_cache[remote] = Ticket(
-                author, remote, end_timestamp  # type: ignore
-            )
+            self.ticket_cache[author] = self.ticket_cache[remote] = {
+                "author": author,
+                "remote": remote,
+                "timestamp": end_timestamp,
+            }
 
     async def handle_ticket_db_side(self, author_id: int, remote_id: int):
         timestamp = discord.utils.utcnow()
         timestamp = timestamp + datetime.timedelta(4)
-        self.ticket_cache[author_id] = self.ticket_cache[remote_id] = Ticket(
-            author_id, remote_id, end_timestamp  # type: ignore
-        )
+        self.ticket_cache[author_id] = self.ticket_cache[remote_id] = {
+            "author": author_id,
+            "remote": remote_id,
+            "timestamp": timestamp,
+        }
         unix = timestamp.timetuple()
         unix = time.mktime(unix) * 1000
-        await self.pool.execute(
-            "INSERT INTO TICKETS VALUES ($1, $2, $3)", author_id, remote_id, unix
-        )
+        await self.pool.execute("INSERT INTO TICKETS VALUES ($1, $2, $3)", author_id, remote_id, unix)
 
     @commands.command(brief="Closes support ticket.")
     async def close(self, context: JDBotContext):
         if not context.guild:
             if context.author.id not in self.ticket_cache:
                 return await context.send("You did not create any tickets.")
-            remote = self.ticket_cache[context.author.id].remote
-            remote_channel = await self.bot.try_channel(remote)
-            await remote_channel.send("The user closed the ticket.")
+            remote = self.ticket_cache[context.author.id]["remote"]
+            remote = self.bot.get_channel(remote)
+            await remote.send("The user closed the ticket.")
 
         else:
             if not context.channel.id in self.ticket_cache:
                 return await context.send("This is not ticket channel.")
-            ticket = self.ticket_cache[context.channel.id]
-            author = ticket.author
-            remote = ticket.remote
-            author = await self.bot.try_user(author)
+            author = self.ticket_cache[context.channel.id]["author"]
+            author = self.bot.get_user(author)
             await author.send("Support team has closed your ticket.")
 
-        del self.ticket_cache[remote]
+        del self.ticket_cache[remote.id]
         del self.ticket_cache[context.author.id]
         await self.pool.execute(
-            "DELETE FROM TICKETS WHERE author_id = $1 AND remote_id = $2",
-            context.author.id,
-            remote,
+            "DELETE FROM TICKETS WHERE author_id = $1 AND remote_id = $2", context.author.id, remote.id
         )
 
-    @commands.command(
-        brief="creates a ticket for support", aliases=["ticket_make", "ticket"]
-    )
+    @commands.command(brief="creates a ticket for support", aliases=["ticket_make", "ticket"])
     @commands.dm_only()
-    async def create_ticket(
-        self, context: JDBotContext, *, starter_message: Optional[str] = None
-    ):
+    async def create_ticket(self, context: JDBotContext, *, starter_message: Optional[str] = None):
         if context.author.id in self.ticket_cache:
-            return await context.send(
-                "You cannot create another ticket while a ticket is not responded."
-            )
+            return await context.send("You cannot create another ticket while a ticket is not responded.")
         ticket_channel = self.main_channel
         thread_channel = await ticket_channel.create_thread(
             name=f"User {context.author} - {context.author.id}",
@@ -128,21 +100,11 @@ class TicketCog(commands.Cog, name="Ticket"):
         await self.handle_ticket_db_side(context.author.id, thread_channel.id)
 
         if not starter_message:
-            starter_message = (
-                "Hello i need help, i Haven't provided a reason quite yet."
-            )
-            modal = utils.ReportView(context, timeout=180.0)
-            message = await context.send(
-                "Please Submit your Reason Here(if you don't respond with in 3 minutes, it will use a default):",
-                view=modal,
-            )
-            await modal.wait()
-            starter_message = modal.value or "I need help, i've provided no reason."
+            starter_message = "Hello i need help, i Haven't provided a reason quite yet."
+            # place holder for now will use modals later
 
         await thread_channel.send(f"`{context.author}:` {starter_message}")
-        await context.send(
-            "Created, now you can keep sending messages here to send it to remote channel."
-        )
+        await context.send("Created, now you can keep sending messages here to send it to remote channel.")
         await ticket_channel.send("<@168422909482762240> New support ticket.")
 
     @create_ticket.error
@@ -172,31 +134,25 @@ class TicketCog(commands.Cog, name="Ticket"):
         # edit later idk
 
         if not self.main_channel:
-            self.main_channel: Any = self.bot.get_channel(855947100730949683)
-            self.support_role = self.bot.get_guild(736422329399246990).get_role(
-                855219483295875072
-            )
+            self.main_channel = self.bot.get_channel(855947100730949683)
+            self.support_role = self.bot.get_guild(736422329399246990).get_role(855219483295875072)
 
-        if (
-            message.guild
-            and message.guild.id == 736422329399246990
-            and message.channel.id in self.ticket_cache
-        ):
+        if message.guild and message.guild.id == 736422329399246990 and message.channel.id in self.ticket_cache:
 
             if self.support_role not in message.author.roles:
                 return await message.reply(
                     "<a:yangsmh:800522615235805204> Sorry you can't use this as you aren't a support team member."
                 )
 
-            author = self.ticket_cache[message.channel.id].author
-            author = await self.bot.try_user(author)
+            author = self.ticket_cache[message.channel.id]["author"]
+            author = self.bot.get_user(author)
             await author.send(f"`{message.author}:` {message.content}")
 
         elif not message.guild and message.author.id in self.ticket_cache:
-            thread = self.ticket_cache[message.author.id].remote
-            thread = await self.bot.try_channel(thread)
+            thread = self.ticket_cache[message.author.id]["remote"]
+            thread = self.bot.get_channel(thread)
             await thread.send(f"`{message.author}:` {message.content}")
 
 
 async def setup(bot: JDBot):
-    await bot.add_cog(TicketCog(bot))
+    await bot.add_cog(Ticket(bot))
